Refactor origami model export.
