+++ src/main/java/oripa/persistent/entity/OrigamiModelFileTypeKey.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.persistent.doc;
+package oripa.persistent.entity;
 
 import oripa.persistent.filetool.Exporter;
 import oripa.persistent.filetool.FileTypeProperty;
